Fix #32 by appending ./ to relative imports from the same folder
When generating from docs using a custom baseQuery which is in the same folder as the output api.generated.ts file the resolveImportPath fails to append ./ to the start of the import path. Most likely because the path.relative call resolves the path to an empty string if the two directories are the same.

This fix that assumes that resolveImportPath is only called for relative imports, and ignores node modules which in my testing seems to be the case.

Written test and helper function to copy customBaseQuery into test/tmp folder and run a generate command using it. All tests pass successfully.